requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distro==1.8.0
dotenv-python==0.0.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
matrix-client==0.4.0
python-dotenv==1.0.0
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
zulip==0.8.2
